Introduce TaskMap abstraction into `tokio-tasks`
